refactor: split lms media modules
