+++ packages/shared/src/lms/media/index.ts
@@ -1,0 +1,4 @@
+export { getLectureTranscript, listLectureTranscripts, createLectureTranscript } from './transcript';
+export { listLectureNotes, createLectureSummaryNote } from './summary';
+export { listAudioExtractions, createAudioExtraction } from './audio';
+export { getLecturePipeline, buildPipelineOverview } from './pipeline';
